Add __playerName variable
Adds the self.__playerName variable in the Player class
which allows the player to name their character. Also
added functionality to get the player's name.

Adds a KeyboardInterrupt exception in the main function
call so that the user can exit the program early and
return with a much cleaner exit message.