Changing grace() method functionality
Changed the functionality of the grace() method for the player
objects by removing the 1st sleep timer. Created a for loop to loop
through each player in the two_player_game and three_player_game
functions to heal each player object and then activate a short sleep
timer. This prevents having to wait the previous 3 seconds for each
player object that was in the game before moving on to the next fight.