Adding summon deaths
Adds the ability for the "summon" characters to reach 0 hp and die
in the boss fight. This allows for the boss to kill the summon
characters and continue fighting as long as the host is still alive. It
also prevents the summon characters from dealing damage to the boss
after they have reached 0 hp.