Adding boss death to end fight
Adds the ability for the boss' hp to reach 0 and end the fight without
allowing the leftover summon players from continuing to attack the boss
after it has already been defeated.

Example: If there is a host and 2 summons, if the boss' hp reaches 0
from an attack by the first summon, then the second summon's attack
phase is skipped.

Also increased the sleep time in the Character grace() method so the
text does not move as fast on the screen when multiple players are
resting.